[AFID-280] Refactored test code in line with review suggestions.]
